requirements.txt
Jinja2==3.1.2
libvirt-python==8.0.0
